# src/main/java/com/example/bankcards/dto/response/JwtResponse.java
package com.example.bankcards.dto.response;

import lombok.Data;

import java.util.List;

@Data
public class JwtResponse {
    private String token;
    private String type = "Bearer";

    public JwtResponse(String token) {
        this.token = token;
    }
}